refactor: moved work page content to artistry page
