+++ app/[lang]/artistry/page.test.tsx
@@ -1,0 +1,28 @@
+import { render, screen } from '@testing-library/react';
+
+import Artistry from './page';
+
+jest.mock('next-intl/server', () => ({
+  getTranslations: jest.fn().mockResolvedValue((key: string) => key)
+}));
+
+jest.mock('~/components/title-with-quote/TitleWithQuote', () => {
+  const MockLiatoshynskyArtistry = () => <div>Liatoshynksy artistry</div>;
+  MockLiatoshynskyArtistry.displayName = 'MockLiatoshynskyArtistry';
+  return MockLiatoshynskyArtistry;
+});
+
+jest.mock('./CompositionTable/MusicTableSelection', () => {
+  const MockCompositionTable = () => <div>Composition table</div>;
+  MockCompositionTable.displayName = 'MockCompositionTable';
+  return MockCompositionTable;
+});
+
+describe('Liatoshynsky artistry page', () => {
+  it('should render Liatoshynsky artistry page correctly', async () => {
+    render(await Artistry());
+
+    expect(screen.getByText(/Liatoshynksy artistry/i)).toBeInTheDocument();
+    expect(screen.getByText(/Composition table/i)).toBeInTheDocument();
+  });
+});
