+++ i18n/messages/en.json
@@ -102,9 +102,9 @@
       }
     }
   },
-  "liatoshynskyWork": {
+  "liatoshynskyArtistry": {
     "title-with-quote": {
-      "title": "LIaTosHynSKy's WoRK",
+      "title": "LIaTosHynSKy's ArTiStRy",
       "quoteText": "Oh, my dear, my dear darling, when will that time finally come when we will be together, in the living room, where it's so beautiful, there's a piano and lots of sheet music.",
       "sourceText": {
         "title": "Letter from Borys Liatoshynsky to Margarita Tsarevych",
