+++ i18n/messages/uk.json
@@ -104,7 +104,7 @@
       }
     }
   },
-  "liatoshynskyWork": {
+  "liatoshynskyArtistry": {
     "title-with-quote": {
       "title": "ТвоРчіСТь ЛЯтоШиНськогО",
       "quoteText": "Ах, мила, милий мій котику, коли ж нарешті прийде той час, коли ми будемо разом із тобою, у вітальні, де так гарно, стоїть рояль і багато нот.",
